eapi.sh: match by extended glob rather than by ERE
As regards the [[ keyword, bash always treats the right-hand side of ==
and != as an extended globbing pattern. However, prior to bash-4.1, the
parser required for the extglob shopt to be enabled in order to tolerate
a raw extglob. Otherwise, the only alternative was to conceal the
extglob by expanding it from a variable.

Given that I am targeting bash-4.2+, these constraints no longer apply.
As such, employ conventional pattern matching throughout eapi.sh.
Testing with == and != is measurably faster than with =~ and is easier
on the eye in most cases.

It should be noted that extglobs support branching, just as regular
expressions do. For instance, @([0-9]|4-slot-abi|10) would be a
perfectly legal pattern.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>