Don't expand parameters as $@ for assignment to scalars
Per SC2124, though bash does promise that this approach results in the
elements being joined by a space, it is often a code smell. Instead,
either explicitly join the positional parameters as "$*" or expand
whichever parameter is relevant.

Also, for some of the functions in bin/phase-helpers.sh, check the
positional parameters with has() first, rather than join and proceed to
check by way of an unquoted variable expansion, which may then result in
unintentional pathname expansion (SC2086).

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>