phase-helpers.sh: don't declare __unpack_tar() repeatedly in loop
Within the unpack() function, the __unpack_tar() function is re-declared
for every single positional parameter that is iterated over.

I have noticed the tendency for functions to define other functions, and
estimate that this strategy is in place for two reasons. Firstly, so as
not to clutter the function namespace with functions that may otherwise
remain unused. Secondly, to rigorously defend against portage-owned
functions being replaced elsewhere, such as an in ebuild or an eclass.

Be that as it may, declaring functions is not cost-free and doing so
within a loop is simply going too far. Let's hoist the function upwards
so that it is declared just after unpack() validates its arguments.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>