phase-helpers.sh: use $# to check the argument count in unpack()
Because that is the proper way to do it. [[ -z "$*" ]] is simply
nonsense. If the intention was to defend against empty arguments then it
does no such thing; such a check would be easily circumvented by
specifying at least two arguments, be they empty or not. It is also
sensitive to the value of IFS.

Besides, if the empty string is encountered then unpacking will fail
anyway. In that event, so be it. Render the diagnostic message
indicating commencement of unpacking more useful by expanding the
filename with ${param@Q} expansion. Such is acceptable, given a target
of bash 4.4.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>