phase-helpers.sh: refrain from meddling with nullglob
Both _eapply_patch() and eapply_user() set nullglob before looking for
patches. However, both also iterate over * before proceeding to compare
the subject variable to the patterns *.diff and *.patch. Given that,
it's not really worth the effort of enabling nullglob. In the case that
the glob expands as itself, it can't possibly match either of those
patterns. In turn, there is no justification for incurring the expense
of a subshell to grab the output of shopt -p.

While at it, use a raw extglob to compare against the aforementioned
patterns, serving as a micro-optimisation. Note that, since bash-4.1,
this is always legal syntax, irrespective of whether the applicable
shopt is enabled or not.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>