phase-helpers.sh: avoid needless repetition of die in unpack()
Instead, check the exit status value propagated by the case/esac
compound command itself.

Simplify the strictures concerning pipeline error detection by
localising the - parameter and arbitrarily enabling the pipefail shell
option. This is acceptable, given a target of bash >=4.4.

Further, execute mv(1) only once in the course of unpacking deb
archives. Presently, ithe routine calls upon deb2targz to unpack deb
archives before proceeding to run the following command.

  mv -f "${basename%.deb}".tar.gz data.tar.gz

If that command fails, it falls back to the following command.

  mv -f "${basename%.deb}".tar.xz data.tar.xz

The approach is somewhat flawed. Just because the first command exits
non-zero, it is not necessarily the case that the error be a consequence
of the file missing (ENOENT). Yet, such an inference is made. Rather
than potentially execute two mv(1) command in succession, ascertain the
name of the file to be renamed by using pathname expansion before
executing mv(1) exactly once.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>