phase-helpers.sh: refactor eapply_user()
The eapply_user() function currently spreads the logic of selecting
which patches to apply across a procedure which is dedicated to
selecting patches and a procedure which applies them. This makes the
code harder to read and reason with than would otherwise be the case.

This commit cleans things up by walking the patch directories in
ascending order of specificity, as opposed to descending. Once all of
the directories have been traversed, the associative array correctly
defines the patches which need to be applied, which was not previously
the case. In turn, this simplifies the process of applying them. Gone is
the "applied" variable and the clumsy code surrounding it. Rather, if
the array is populated, simply get on with it.

The name of the array has been changed. The name, _eapply_user_patches,
looked too much like the name of a non-public function. Besides, the
underscore doesn't make a great deal of sense; there is no question of
its provenance as a local variable and none of the other local variables
have them.

Don't compose a horizontal rule by employing a needless loop and
subshell, twice.

Set IFS= while reading in the basenames. Otherwise, the basenames will
not necessarily be conveyed verbatim, resulting in an obscure bug
whereby a patch whose basename is only distinguished from another by
leading whitespace would erroneously be considered as the same patch.

Remove the superfluous -- operand from printf.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>