phase-helpers.sh: build decompression command from __unpack_tar() args
This does away with the unquoted expansion of $1 but there is more to it
than merely placating shellcheck. Before, it was possible for a
user-specified command to induce pathname expansion, for which there is
no compelling use case. Now, it is impossible for this to occur. For
example, specifying PORTAGE_BUNZIP2_COMMAND="/bin/*" will now result in
attempting to execute a binary by that name, as opposed to expanding all
the files in /bin as distinct words then trying to execute a simple
command beginning with whatever the first of those happens to be.

Further, given a blank command - such as PORTAGE_BUNZIP2_COMMAND=" " -
portage would previously attempt to execute a command named -c. Now, it
will instead fall back to using PORTAGE_BZIP2_COMMAND. Should that also
yield a blank command, an exception shall be thrown.

It should be noted that the act of localising IFS is sufficient to
guarantee the default behaviour, provided that no assignment occurs.
That is, word splitting shall be performed as if a value of $' \t\n'
were in effect.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>