phase-helpers.sh: improve diagnostics for unpackable files
Just prior to the point at which unpacking commences, an -s test is
performed.  However, a file could exist while being zero bytes in size,
rendering the error message potentially confusing.

Improve it by first determining whether the file exists and is a regular
file (or a valid symlink to one). If it is not, die with an unambiguous
error message. Likewise, if the file is then found to be empty. Ideally,
one would not bother with testing the size but there may have been a
reason for doing so, such as a poorly implemented archiving utility.
Given that, let's leave it in for now.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>