install-qa-check.d: migrate xdg-utils checks over from preinst/postinst to ${D}
It's practically criminal to run these at merge time instead of src_install
time. It disproportionately affects binpkg consumers, because it applies
to the entirety of ROOT. We actually don't want to do... basically any
of this. It's not even accurate because we are heavily reliant on mtime
of installed files to check whether the commands were actually run.

What we actually want to do is significantly simpler: every package that
installs specific files to ${D} has to also inherit an eclass and run a
function in pkg_postinst. We can check for inherits quite trivially, and
warn about those. We can also slightly less efficiently check the
contents of pkg_* functions to see if they make certain calls; bash can
print the function contents and we can grep for that.

It doesn't catch cases where a custom eclass runs the xdg functions, but
we manually include those in our matching.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>