Pass through MAKEFLAGS and exclude from environment.bz2
Allow the MAKEFLAGS environment variable to pass through, in case a
centralized GNU Make POSIX Jobserver is available. In order to prevent
persistence of this variable in environment.bz2, exclude it when the
__save_ebuild_env --exclude-init-phases argument is given.

Ultimately we may want to add support for portage to parse MAKEFLAGS
and use it to allocate job tokens in various circumstances. For
example, emerge could allocate a job token for each job started for
emerge --jobs. This would remove a job token from the pool that
is available to nested make calls, but is reasonable because nested
make calls will execute jobs serially when no jobserver tokens remain.

Bug: https://bugs.gentoo.org/692576
Signed-off-by: Zac Medico <zmedico@gentoo.org>