sync: add a 5 second timeout to gemato's openpgp refresh
If it takes more than 5 seconds to download a miniscule file from
https://gentoo.org/.well-known/openpgpkey/ then something has gone
dreadfully wrong. Most commonly, what went wrong is that the user has
broken ipv6 connectivity and requests simply hangs forever (no joke --
it doesn't implement Happy Eyeballs and closed the multiple bug reports
as "please use stackoverflow to ask questions about how to use requests,
we are not a support forum").

Pass a timeout so that we eventually give up and try ipv4 instead. This
is a crude hack and the proper solution is to make gemato handle this
better, but until gemato is fixed to use a better download library we do
the best we can.

Bug: https://github.com/psf/requests/issues/6788
Bug: https://github.com/projg2/gemato/issues/35
Bug: https://bugs.gentoo.org/779766
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>