Drop _emerge/getloadavg.py
Portage's custom getloadavg.py wrapper was added with
8607a39dc9d9 ("If necessary, use /proc/loadavg to emulate
os.getloadavg()."). All platforms we care about provide
os.getloadavg() now, so this can be dropped.

Zac did a nice analysis of the historical background in

https://github.com/gentoo/portage/pull/1383#issuecomment-2375471855

to summarize: python on uclibc did not had os.getloadavg() in the
past, but uclibc (and uclibc-ng) where retired in gentoo [1].

1: https://www.gentoo.org/support/news-items/2021-08-18-uclibc-ng-retirement.html
Signed-off-by: Florian Schmaus <flow@gentoo.org>