ebuild.sh: Allow CC and CXX to be changed in profile.bashrc
Code added in 2003 previously reset these to their original values after
sourcing all the bashrc files. It isn't clear, but this was possibly
done to allow the user's values to take precedence over the old
profile.env. Today, only the llvm profiles set CC/CXX and this is done
in make.defaults, which doesn't override the user's values. That makes
this code redundant, but it would be useful to conditionally adjust
CC/CXX in llvm's profile.bashrc when cross-compiling.

Signed-off-by: James Le Cuirot <chewi@gentoo.org>