Add PORTAGE_LOG_FILE_SEP option
Portage uses the colon (:) as a separator between the components of a
log file under PORTAGE_LOGDIR. However, the colon is often used as a
"word" separator by shells and consoles, which makes it not
user-friendly to select the full path of the log file. For example,
double-clicking on a portage log file path in Kitty will not select
the complete path, only the substring of the path up to the nearest
colon.

Therefore, this introduces PORTAGE_LOG_FILE_SEP, which allows the user
to choose the separator used for log files. To keep the existing
behavior, PORTAGE_LOG_FILE_SEP defaults to ':'.

Signed-off-by: Florian Schmaus <flow@gentoo.org>