estrip: salt build IDs with ${CATEGORY}/${PF}:${SLOT}
Use debugedit to salt build IDs with ${CATEGORY}/${PF}:${SLOT} to avoid
file collisions (same build ID generated as another package) where a bundled
library is installed.

This is initially done only with FEATURES=installsources as that's the
case we require debugedit in.

Bug: https://bugs.gentoo.org/549672
Bug: https://bugs.gentoo.org/953869
Signed-off-by: Sam James <sam@gentoo.org>
Part-of: https://github.com/gentoo/portage/pull/1436
Signed-off-by: Sam James <sam@gentoo.org>