estrip: salt build IDs even with FEATURES=-installsources
We've historically only needed debugedit for FEATURES=installsources
but if we're going to enable build IDs, we need to salt them in general
for splitdebug.

This is done in a followup commit to the original salting as the
implementation could be cleaner and I think it's easier to rework then
if required.

Bug: https://bugs.gentoo.org/549672
Bug: https://bugs.gentoo.org/953869
Signed-off-by: Sam James <sam@gentoo.org>