90gcc-warnings: employ LC_COLLATE and LC_CTYPE consistently
Presently, LC_ALL=C is set for the invocation of sed(1), whereas
LC_COLLATE and LC_CTYPE are set for the invocation of grep(1). This is
inconsistent in so far as LC_ALL is permitted to override the intended
character type and collation, yet only in the case of the latter. Given
that the processing of the log occurs in a subshell, render things
consistent by exporting all of LC_ALL, LC_COLLATE and LC_CTYPE.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>