90gcc-warnings: convey warnings in a memory-efficient manner
Rather than synchronously slurp all of the warnings matched against the
contents of PORTAGE_LOG_FILE into a scalar variable before displaying,
read the warnings line by line while asynchronously calling eqawarn for
each. This may help to save memory in the case that many warnings are to
be displayed.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>