phase-helpers.sh: name and employ unpack() variables coherently
This commit accomplishes the following:-

- introduces a new 'inner_suffix' variable
- introduces a new 'basename' variable
- renames the 'x' variable to 'f'
- drops the 'cwd_dest' variable
- drops the 'y' variable

The rationale for these changes is described herewith.

Firstly, 'f' is slightly more intuitive as the name of a variable whose
value refers to a regular file (or a symlink to one). While 'x' might
commonly be inferred as an item collected from the positional parameter
list, it affords the reader no immediate insight as to its provenance.

Secondly, the basename of 'f' is considered not only within the body of
the loop that iterates over the positional parameters, but also within
the body of the __unpack_tar() helper function. In the former case, it
is calculated while handling deb archives, just before clobbering the
value of 'y'. In the latter case, it is calculated and assigned to a
variable named 'cwd_dest'. Instead, simply assign the basename to the
rather appropriately named 'basename' variable.

Thirdly, drop the 'y' variable. Not only does its name convey no
semantic meaning, it was being employed for two starkly contrasting
purposes:

1) holding the value of the inner suffix e.g. "f.tar.gz" -> "tar"
2) holding the basename of 'f'

As regards the first case, the __unpack_tar() now defines a localised
'inner_suffix' variable. As regards the second case, it is now covered
by the 'basename' variable.

I am of the view that these otherwise unobtrusive changes render the
intended behaviour of the unpack() function markedly more apparent,
especially for those who may be reading the code for the first time.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>