phase-helpers.sh: use hash to gauge deb2targz availability
Presently, the unpack() function uses the type builtin to determine
whether deb2targz is present in PATH.

type -P deb2targz >/dev/null # proves that deb2targz is in PATH
deb2targz "${basename}"      # deb2targz is then invoked in this manner

However, using type -P in this way is questionable. While type -P does
coerce bash into performing a PATH search, the manner in which deb2targz
is invoked still permits for deb2targz to be a function - even an alias
in the case that the expand_aliases option is enabled. Put another way,
type -P is worth using where the pathname of an executable must be
determined and subsequently referenced. Given that no such requirement
is implied, use the hash builtin to determine whether deb2targz is
available as a command. This has the advantage of pre-warming the cache
that the shell employs for Command Search and Execution.

Also, perform the command search before potentially performing a command
substitution for ar --version. The latter operation will almost
certainly be more expensive than the former.

It should be noted that the revised approach incurs fewer dup2(2) calls,
owing to STDERR being redirected just once.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>