phase-helpers.sh: refine the implementation of use()
Refrain from assigning a value to IFS in the course of localising it. To
localise IFS without assignment incurs the behaviour associated with IFS
being unset, which is wholly adequate.

Jettison the 'prev_shopts' and 'ret' variables. Instead, localise the -
parameter. Such is acceptable, given a target of >=bash-4.4.

Employ pattern matching to determine whether the flag begins with "!".

Rename the 'found' variable to 'invert'.

Use declare -F to probe for the existence of the ___in_portage_iuse
function so that no time is wasted printing out the entirety of the
function's body to /dev/null.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>