phase-helpers.sh: don't wordsplit the given flag in use()
Presently, the use() function splits the value of $u in the course of
conveying it to the has() function. This is a rather serious bug that
would have been more serious still, were it not for the fact that a
check against IUSE is performed in most cases. Still, consider what can
happen in the case that the check is not performed.

use "baboon baboon"                  # will always return true
use "baboons write bash for baboons" # ditto

The solution is to quote the expansion of $u, thereby protecting it from
wordsplitting.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>