ebuild.sh: avoid arbitrarily disabling extglob
There are several instances in which a dynamic function declaration
could theoretically be interpreted as an extglob, depending on the exact
name of the function. Presently, the defence against this is to forcibly
disable extglob - an option that is best left untouched except where
being wilfully enabled as soon as bash starts. Instead, place a space
between the name of the function and the following pair of parentheses.

I must add that bash never enables the extglob option unless it is
directly instructed to do so.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>