misc-functions.sh: avoid arbitrarily toggling nullglob
The nullglob option is being disabled by install_hooks(), regardless of
whether it was enabled prior to invoking the function. The simplest
solution is to not set nullglob. The subsequent -x test will already
fail in the case that the glob expands as it itself without "*" being
the actual name of an executable file.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>