phase-helpers.sh: name and employ variables sensibly in eapply()
In the course of parsing the arguments conveyed to the eapply()
function, add options to an array named 'options', and operands to an
array named 'operands'.

Employ a variable named 'path' to iterate over the members of
'operands', rather than repurpose the 'i' variable. Doing so makes it
clear that each operand is expected to be a pathname.

Within the _eapply_patch() function, don't rely on the initial set of
patch(1) options being the subject of a variable defined by an upper
scope. Instead, convey the options to the function as positional
parameters.

Within the _eapply_patch() function, don't manipulate a variable defined
by an upper scope. Instead, define a local 'patch_opts' array in which
to incorporate - and augment - the provided patch(1) options.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>