phase-helpers.sh: use hash to gauge gfind(1) availability
Presently, the eapply() function uses the type builtin to determine
whether gfind(1) is present in PATH. To do so is questionable. While
type -P does coerce bash into performing a PATH search, the manner in
which gfind is invoked still permits for it to be a function - even an
alias in the case that the expand_aliases option is enabled. Instead,
use the hash builtin to determine whether gfind is available as a
command. This has the advantage of pre-warming the cache that the shell
employs for Command Search and Execution.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>