phase-helpers.sh: render the exit status of _eapply_patch() meaningful
It isn't necessary to know the exact exit status value of the patch(1)
utility. PMS only requires for eapply() to return a non-zero status in
the case that something goes wrong and nonfatal is in effect.

Even if that were not the case, having _eapply_patch() be implemented as
an impure function that sets a variable in an upper scope just to convey
an exit status value is a rather odd way of going about things.

Simplify all of this by ensuring that _eapply_patch() returns non-zero
in the nonfatal case. Note that it takes advantange of the preceding
commit to do so; it is now guaranteed that __helpers_die() returns
something other than 0, where nonfatal is in effect. Consequently,
eapply() can now respond directly to the return value, rather than
proceed to consult the 'failed' variable, which has been eliminated.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>