Use contains_word() almost everywhere that it is applicable
As was mentioned by the commit introducing the contains_word() function,
it addresses a use case for which the has() function is rampantly
misappropriated. To briefly recap here, has() compares its first operand
to the following operands in a stringwise fashion. However, it is
commonly employed in the following fashion.

  # Here, wordlist contains zero or more whitespace-separated words.
  has some-word ${wordlist}

Thus, the shell is tasked with splitting the expansion of the variable -
a process which is sensitive to the prevailing value of IFS. Further, it
is exceedingly rare for such expansions to be guarded by the "noglob"
shell option, meaning that unintentional pathname expansion remains an
attendant risk.

The contains_word() function safely and efficiently addresses this
particular use case.

  # Unaffected by IFS. Immune to pathname expansion. Fast.
  contains_word some-word "${wordlist}"

This commit replaces the use of has() with contains_word() in almost
every case where it is applicable, though I elected to defer in several
instances that require further investigation. I would add that this
commit touches 18 files which previously accounted for no fewer than 175
SC2086 warnings. This number is thus reduced to 97 in one fell swoop.

See-also: 4a4631eef7186c29668a8c049d988b61469940fd
Link: https://github.com/gentoo/portage/pull/458
Link: https://www.shellcheck.net/wiki/SC2086
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>