phase-helpers.sh: don't perform a dry run in _eapply_patch()
Presently, the _eapply_patch() function executes a dry run of patch(1)
to determine whether the patch can be applied without fuzz. It then
proceeds to execute patch(1) for application, instructing the utility to
be silent in the case that a fuzz factor of zero is permissible.

Going about it in this way is wasteful. Instead, execute patch(1) just
once while capturing its output. In the case that the patch is applied
successfully, convey the output to STDOUT only if it matches a pattern
identifying a fuzz factor. Otherwise, if the patch was unsuccessfully
applied, convey the output to STDERR.

Also, localise IFS so as to guarantee that the positional parameters are
joined by a space in the diagnostic message.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>