phase-helpers.sh: refactor the einstalldocs() function
Declare the entire function as a compound command of the ( … ) form,
which induces a subshell.

Rename the d variable to f, since it is being employed with the
expectation that its value will be the pathname of a regular file.

Use the -v test operator to determine whether the DOCS variable is set.
This feature was introduced by bash-4.2.

Refrain from using the ___is_indexed_array_var() function to determine
whether variables are indexed array variables. Instead, use the
${param@a} form of expansion. This feature was introduced by bash-4.4.

Reduce the use of the && control operator so as to clarify the overall
flow control.

In the course of testing for a scalar variable, test not only that the
variable isn't an indexed array variable, but also that it isn't an
associative array.

Suppress SC2086 warnings where word splitting is being wilfully performed.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>