isolated-functions.sh: improve the ERE detecting -j in MAKEOPTS
In many ways, to parse -j and/or --jobs from MAKEOPTS is a fool's
errand. There is no way of going about it correctly, short of
implementing an option parser that behaves precisely as does that of
make(1) itself.

Still, the regular expression that is presently employed can be made
less worse, so to speak. This commit revises it to be as follows.

.*[[:space:]]
(
        # short options other than -j, if any, leading up to -j,
        # optionally followed by whitespace
        -[^j]*j[[:space:]]*
        |
        # --jobs followed by either = or whitespace
        --jobs(=|[[:space:]]+)
)
([0-9]+)    # the job count ...
[[:space:]] # which must not have any trailing garbage

The following table shows the increased strictness of this expression.

SUBSTRING  BEFORE   AFTER
--jobs= 4  Valid    Invalid
--jobs 4x  Valid    Invalid
-j4x       Valid    Invalid
-jj4       Valid    Invalid

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>