estrip: jettison the debugedit_{found,warned} and dwz_warned variables
Rename the global 'debugedit' variable to 'debugedit_bin', and the
global 'dwz' variable to 'dwz_bin', so as to better convey their
respective purposes.

Jettison the 'debugedit_warned' and 'dwz_warned' variables. Instead,
unset the 'debugedit_bin' and 'dwz_bin' variables to indicate that no
further warnings should be issued.

Jettison the useless 'debugedit_found' variable. For the 'debugedit_bin'
variable to be non-empty guarantees that it was found.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>