estrip: map FEATURES and PORTAGE_RESTRICT to associative arrays
Presently, estrip employs the abhorrent methodology of utilising eval to
dynamically declare variables that are subsequently treated as simple
commands in order to determine whether a given feature or restriction is
in effect, without having to repeatedly scan the FEATURES and
PORTAGE_RESTRICT variables.

Instead, compose two associative arrays, named 'has_feature' and
'has_restriction', whose keys are the features and restrictions of
interest and whose values are either 0 or 1. That makes it possible to
write tests in the following fashion, which I would submit as being
aesthetically pleasing.

(( ! has_feature[dedupdebug] || has_restriction[dedupdebug] ))

Going about it in this way also confers the advantage of eliminating a
slew of SC2154 warnings in shellcheck.

Finally, use the recently added contains_word() function to scan the
FEATURES and PORTAGE_RESTRICT variables. Not only is it faster but it
eliminates several defects pertaining to SC2086 into the bargain. It is
worth noting that this is the first ever instance of the aforementioned
function being utilised by the portage codebase at large.

See-also: 4a4631eef7186c29668a8c049d988b61469940fd
Link: https://github.com/gentoo/portage/pull/458
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>