Don't report a phantom -p1 option in _eapply_patch() exceptions
In the course of recently re-factoring the eapply() and _eapply_patch()
functions, one of the changes made was to have the patch(1) arguments be
reported exactly as they are in the case that an exception is thrown.
However, I forgot to remove a hard-coded instance of "-p1" in the
message template, which is harmless but confusing. Remove it.

Also, use the ${param@Q} form of expansion to convey both the arguments
and the name of the patch. Doing so is acceptable, given a target of
>=bash-4.4.

Fixes: d11490f9d83baefc1dad02d9be63c74e4c1da04f
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>