estrip: track issued warnings context-sensitively
Presently, estrip will raise warnings in the event that is tasked with
doing something that requires for a given tool to be available, yet is
unable to find it. It endeavours never to print the same warning twice
and accomplishes this by unsetting the applicable key of the 'path_of'
map, whose empty value will have conveyed the absence of the tool.

Such is all well and good. However, my attention was drawn to a
still-open GitHub PR which demonstrates a need for tracking whether
warnings have already been issued for a tool in a given context. For
example, there may be a need to warn of the absence of debugedit(1) on
account of needing it to fulfil the behaviour of the "splitdebug"
feature, whereas there may also be a need to warn on account of needing
to fulfil the behaviour of the "installsources" feature.

One way to address this would to go back to using individual scalar
variables to track whether a given warning has been issued. Rather than
do that, this commit introduces an associative array by the name of
'warned_for', whose keys are expected to be the name of a given tool,
and its values, strings comprised by space-separated warning categories.
The code beneath demonstrates how it is intended to be used in practice.

if [[ ! ${path_of[brush]} ]]; then
    # The brush tool is missing but needed for detangling.
    if ! contains_word "hair-detangle" "${warned_for[brush]}"; then
        # No warning yet issued. Go ahead and issue one now.
        warned_for[brush]+=" hair-detangle"
        ewarn "The hair-detangle feature requires the brush tool"
    fi
    return 0
fi

Link: https://github.com/gentoo/portage/pull/1436
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>