estrip: reliably detect missing tools
Presently, estrip performs a PATH search for tools such as strip(1) and
objcopy(1), yet only in the case of their CHOST-prefixed names. Should
this initial PATH search fail, the tool will be presumed to exist by its
non-prefixed name, without going so far as to check whether it actually
does. Consider the following scenario.

1) the "type -P x86_64-pc-linux-gnu-dwz" command exits with 1
2) the tool is thus assumed to be available as "dwz" (without checking)

Though I recently reworked the code, it was not I that introduced this
bug. Rather, I inadvertently preserved it. However, now that I have
identified it, I can certainly fix it. Do so by assigning nothing but
the captured output of the type builtin as values of the 'path_of' map,
irrespective of the name being searched for.

Fixes: a67dc215adb52392505d03870c236e828dca3625
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>