estrip: minor refactoring in advance of an incoming fix
Recently, I adjusted the manner in which estrip determines external
tools to be missing. In doing so, I introduced a problem that needs to
be addressed, which I shall do in short order. The purpose of this
commit is to make some alterations that have no impact whatsoever on the
present behaviour, but will render the upcoming commit less noisy than
it otherwise would have been. Said alterations are as follows.

- rename the 'path_of' variable to 'name_of' (to be explained later)
- use an existing loop to look up the location of dwz(1)
- hoist the debugedit fallback code upwards, to be closer to said loop

Again, no functional changes are made as a result of this particular commit.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>